refactor toolbar, split SizePicker & Helper
