+++ src/pages/Editor/Editor.tsx
@@ -1,14 +1,12 @@
 import { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
-import Swal from 'sweetalert2';
 import { v4 as uuidv4 } from 'uuid';
-import autoVerse from '../../assets/autoVerse.gif';
-import backLinks from '../../assets/backLinks.gif';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { addNote, editNote, selectProfile } from '../../redux/loginSlice';
 import { NoteType } from '../../redux/types';
 import * as styles from '../../utils/styles';
 import CommentBox from './CommentBox';
+import EditorHelper from './EditorHelper';
 import './styles.css';
 
 const category = [
@@ -28,7 +26,6 @@ const Editor = () => {
   const dispatch = useAppDispatch();
   const profile = useAppSelector(selectProfile);
   const [isEdit, setIsEdit] = useState(false);
-  const [toggleHelp, settoggleHelp] = useState(false);
   const [note, setNote] = useState<NoteType>({
     id: uuidv4(),
     title: '',
@@ -138,72 +135,7 @@ const Editor = () => {
         >
           取消
         </div>
-        {/* helper */}
-        <div
-          className={`relative ${styles.themeFlex} w-20 border-l border-y ${
-            toggleHelp
-              ? 'bg-stone-100'
-              : 'bg-stone-300 border-b-2 hover:bg-stone-200 hover:border-b-4 hover:border-l-2'
-          }`}
-        >
-          <span
-            className="cursor-pointer material-symbols-outlined text-lg grow text-center"
-            onClick={() => {
-              settoggleHelp((prev) => !prev);
-            }}
-          >
-            help
-          </span>
-          <div
-            className={`${toggleHelp ? 'flex' : 'hidden'} ${
-              styles.theme
-            } text-lg overflow-y-auto flex-col border absolute z-10 right-0 bottom-[38px] h-80 w-72 p-4 rounded-t-3xl rounded-bl-3xl drop-shadow-xl`}
-          >
-            <div className="mb-3 font-semibold">自動經文輸入</div>
-            <span>
-              可以透過
-              <span
-                className="cursor-pointer text-violet-400"
-                onClick={() =>
-                  Swal.fire({
-                    imageUrl: autoVerse,
-                    imageWidth: 1000,
-                    imageAlt: 'autoVerseImg',
-                    showConfirmButton: false,
-                    background: 'rgba(255,255,255,0)',
-                  })
-                }
-              >
-                經文輸入
-              </span>
-              功能自動輸入經文，只要簡單輸入書卷名稱，章節和經節(
-              <i>例:出埃及記16:4</i>
-              )，再按下<i>空白鍵</i>
-              ，經文就會自動出現。此功能支援書卷簡稱，詳細書卷簡稱可至
-              <i>教學筆記</i>查看。
-            </span>
-            <div className="mt-5 mb-3 font-semibold">建立連結</div>
-            <span>
-              可以透過輸入<i>#</i>後，選取<i>筆記標題</i>
-              (亦可接著輸入搜尋文字)來建立
-              <span
-                className="cursor-pointer text-violet-400"
-                onClick={() =>
-                  Swal.fire({
-                    imageUrl: backLinks,
-                    imageWidth: 1000,
-                    imageAlt: 'backLinksImg',
-                    showConfirmButton: false,
-                    background: 'rgba(255,255,255,0)',
-                  })
-                }
-              >
-                筆記連結
-              </span>
-              ，建立連結後會自動在對應筆記的引用列表新增當前筆記。
-            </span>
-          </div>
-        </div>
+        <EditorHelper />
       </div>
     </div>
   );
